# openapi.yaml
openapi: 3.0.3
info:
    title: MechHub Miniback API
    version: 1.0.0
    description: Standalone backend for MechHub. Authentication is session-based and uses the `mechhub.sid` cookie.
servers:
    - url: http://localhost:3001
      description: Local development server
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: mechhub.sid
    schemas:
...
        ErrorResponse:
            type: object
            required:
                - message
            properties:
                message:
                    type: string
                    example: Not authenticated
...
        SoloChatConversation:
            type: object
            required:
                - id
                - title
                - updatedAt
            properties:
                id:
                    type: integer
                    example: 1
                title:
                    type: string
                    example: Force Balance Basics
                updatedAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:30:00.123Z
        RenameSoloChatConversationRequest:
            type: object
            required:
                - title
            properties:
                title:
                    type: string
                    maxLength: 120
                    example: Statics Week 3 Review
        SoloChatMessage:
            type: object
            required:
                - id
                - conversationId
                - role
                - content
                - status
                - attachments
                - createdAt
                - updatedAt
            properties:
                id:
                    type: integer
                    example: 12
                conversationId:
                    type: integer
                    example: 1
                role:
                    type: string
                    enum: [user, assistant]
                    example: assistant
                content:
                    type: string
                    example: Start by drawing the free-body diagram.
                status:
                    type: string
                    enum: [streaming, completed, failed]
                    example: completed
                attachments:
                    type: array
                    items:
                        $ref: "#/components/schemas/SoloChatImageAttachment"
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:31:00.456Z
                updatedAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:31:02.012Z
        CreateSoloChatMessageRequest:
            type: object
            properties:
                content:
                    type: string
                    example: Explain free-body diagrams.
                attachmentIds:
                    type: array
                    items:
                        type: integer
                    example: [7, 8, 9]
        SoloChatImageAttachment:
            type: object
            required:
                - id
                - kind
                - purpose
                - fileName
                - url
                - mimeType
                - width
                - height
                - sizeBytes
                - createdAt
            properties:
                id:
                    type: integer
                    example: 7
                kind:
                    type: string
                    enum: [image, document]
                    example: image
                purpose:
                    type: string
                    enum: [solochat]
                    example: solochat
                fileName:
                    type: string
                    example: beam-notes.md
                url:
                    type: string
                    example: /uploads/files/7
                mimeType:
                    type: string
                    example: image/webp
                width:
                    type: integer
                    nullable: true
                    example: 1536
                height:
                    type: integer
                    nullable: true
                    example: 1024
                sizeBytes:
                    type: integer
                    example: 482913
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-25T09:31:00.456Z
paths:
...
    /uploads/files/{fileId}:
        get:
            tags:
                - Uploads
            summary: Download a previously uploaded file attachment
            operationId: getUploadedFile
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: fileId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Uploaded file
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                format: binary
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: File not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            tags:
                - Uploads
            summary: Delete an unattached uploaded file
            operationId: deleteUploadedFile
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: fileId
                  required: true
                  schema:
                      type: integer
            responses:
                "204":
                    description: Uploaded file deleted
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: File not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: File is already attached to an assignment or submission
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /solochat/conversations:
        get:
            tags:
                - SoloChat
            summary: Get the current user's SoloChat conversations
            operationId: listSoloChatConversations
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: SoloChat conversations ordered by latest activity
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/SoloChatConversation"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        post:
            tags:
                - SoloChat
            summary: Create a new empty SoloChat conversation
            operationId: createSoloChatConversation
            security:
                - cookieAuth: []
            responses:
                "201":
                    description: Conversation created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SoloChatConversation"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /solochat/conversations/{conversationId}:
        patch:
            tags:
                - SoloChat
            summary: Rename a SoloChat conversation
            operationId: renameSoloChatConversation
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: conversationId
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RenameSoloChatConversationRequest"
            responses:
                "200":
                    description: Conversation renamed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SoloChatConversation"
                "400":
                    description: Invalid conversation id or title
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            tags:
                - SoloChat
            summary: Delete a SoloChat conversation
            operationId: deleteSoloChatConversation
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: conversationId
                  required: true
                  schema:
                      type: integer
            responses:
                "204":
                    description: Conversation deleted
                "400":
                    description: Invalid conversation id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /solochat/conversations/{conversationId}/messages:
        get:
            tags:
                - SoloChat
            summary: Get all messages for a SoloChat conversation
            operationId: listSoloChatMessages
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: conversationId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Conversation messages
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/SoloChatMessage"
                "400":
                    description: Invalid conversation id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /solochat/conversations/{conversationId}/messages/stream:
        post:
            tags:
                - SoloChat
            summary: Store the user message and stream the assistant reply
            description: >
                Persists the user message immediately, creates an assistant placeholder
                message with `status=streaming`, then streams newline-delimited JSON
                events while the AI reply is generated. The user message is not returned
                in the stream response.
            operationId: streamSoloChatMessage
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: conversationId
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateSoloChatMessageRequest"
            responses:
                "200":
                    description: NDJSON stream of assistant events
                    content:
                        application/x-ndjson:
                            schema:
                                type: string
                                example: |
                                    {"type":"assistant_start","message":{"id":2,"conversationId":1,"role":"assistant","content":"","status":"streaming","attachments":[],"createdAt":"2026-03-23T09:31:00.500Z","updatedAt":"2026-03-23T09:31:00.500Z"}}
                                    {"type":"assistant_delta","messageId":2,"delta":"Start by drawing "}
                                    {"type":"assistant_delta","messageId":2,"delta":"the free-body diagram."}
                                    {"type":"conversation_title","conversation":{"id":1,"title":"Free-Body Diagram Basics","updatedAt":"2026-03-23T09:31:02.100Z"}}
                                    {"type":"assistant_done","message":{"id":2,"conversationId":1,"role":"assistant","content":"Start by drawing the free-body diagram.","status":"completed","attachments":[],"createdAt":"2026-03-23T09:31:00.500Z","updatedAt":"2026-03-23T09:31:02.050Z"}}
                "400":
                    description: Invalid conversation id or message content
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error or missing AI configuration before the stream starts
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"